Generate 16x16 grid of divs
Added createSquareDiv() which creates a div with a width and heigth
based on the dimensions of the client page. Added createGrid() which
creates a 16x16 grid of divs. Added event listener on the window
that resizes each div if the window is being resized.

The HTML file just contains a container that is being populated by the
grid of divs and the CSS is applying a display flex, so the divs wrap.